Unify ObjID types across API and ABI crates.
