+++ src/lib/twizzler-abi/src/runtime/object.rs
@@ -61,8 +61,7 @@ impl ObjectRuntime for MinimalRuntime {
         flags: twizzler_runtime_api::MapFlags,
     ) -> Result<twizzler_runtime_api::ObjectHandle, twizzler_runtime_api::MapError> {
         let slot = global_allocate().ok_or(MapError::OutOfResources)?;
-        let _ = sys_object_map(None, ObjID::new(id), slot, flags.into(), flags.into())
-            .map_err(|e| e.into())?;
+        let _ = sys_object_map(None, id, slot, flags.into(), flags.into()).map_err(|e| e.into())?;
         Ok(ObjectHandle::new(
             NonNull::new(Box::into_raw(Box::new(InternalHandleRefs::default()))).unwrap(),
             id,
