+++ src/runtime/dynlink/src/engines/twizzler.rs
@@ -107,7 +107,7 @@ impl ContextEngine for Engine {
             }
 
             Ok(ObjectSource::new_copy(
-                twizzler_object::ObjID::new(src.obj.id),
+                src.obj.id,
                 (src_start % MAX_SIZE) as u64,
                 (dest_start % MAX_SIZE) as u64,
                 len,
@@ -136,7 +136,7 @@ impl ContextEngine for Engine {
             .map_two_objects(
                 text_id,
                 MapFlags::READ | MapFlags::EXEC,
-                data_id.as_u128(),
+                data_id,
                 MapFlags::READ | MapFlags::WRITE,
             )
             .map_err(|_| DynlinkErrorKind::NewBackingFail)?;
